# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Home Sale Helper API
  version: "0.1.0"
  description: |
    REST API for the Home Sale Helper micro-SaaS.
    Handles listings, AI generation, photo upload, cross-platform publishing, messaging, and subscriptions.
  contact:
    name: Backend Developer
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.homesalehelper.app
    description: Production (Railway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /platform-accounts/oauth/ebay/callback:
    post:
      summary: Exchange eBay auth code for tokens
      tags: [Platform Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        "201":
          description: Account connected
...
  /platform-accounts/oauth/facebook/callback:
    post:
      summary: Exchange Facebook auth code for long-lived token
      tags: [Platform Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "201":
          description: Account connected
  /platform-accounts/{platform}:
    parameters:
      - name: platform
        in: path
        required: true
        schema:
          type: string
          enum: [EBAY, FACEBOOK, POSHMARK, OFFERUP]
    delete:
      summary: Disconnect a marketplace account
      tags: [Platform Accounts]
      responses:
        "204":
          description: Disconnected

  # ── Messages ──────────────────────────────────────────────────────────────
...